Split BibleReaderViewModelTests into smaller files, address PR comments
